+++ finagle-base-http/src/test/scala/com/twitter/finagle/http/filter/DtabFilterTest.scala
@@ -99,4 +99,23 @@ class DtabFilterTest extends AnyFunSuite with AssertionsForJUnit {
       assert(receivedDtab == Some(origDtab))
     }
   }
+
+  test("Injector does not transmit dtab.limited") {
+    var receivedDtab: Option[Dtab] = None
+    val svc = new DtabFilter.Injector().andThen(Service.mk[Request, Response] { req =>
+      receivedDtab = HttpDtab.read(req).toOption
+      Future.value(Response())
+    })
+
+    Dtab.unwind {
+      val newDtab = Dtab.read("/s => /srv/smf1")
+      Dtab.limited = newDtab
+
+      // prepare a request and add a correct looking new-style dtab header
+      val req = Request()
+      Await.result(svc(req), timeout)
+
+      assert(receivedDtab == Some(Dtab.empty))
+    }
+  }
 }