+++ finagle-core/src/main/scala/com/twitter/finagle/naming/BindingFactory.scala
@@ -95,9 +95,9 @@ class BindingFactory[Req, Rep] private[naming] (
 
   private[this] val dtabCache = {
     val newFactory: Dtabs => ServiceFactory[Req, Rep] = {
-      case Dtabs(baseDtab, localDtab) =>
+      case Dtabs(baseDtab, localDtab, limitedDtab) =>
         val dynFactory = new DynNameFactory(
-          NameInterpreter.bind(baseDtab ++ localDtab, path),
+          NameInterpreter.bind(baseDtab ++ limitedDtab ++ localDtab, path),
           nameTreeCache,
           statsReceiver = statsReceiver
         )
@@ -124,7 +124,8 @@ class BindingFactory[Req, Rep] private[naming] (
               // we don't have the dtabs handy at the point we throw
               // the exception; fill them in on the way out
               case e: NoBrokersAvailableException =>
-                Future.exception(new NoBrokersAvailableException(e.name, baseDtab, localDtab))
+                Future.exception(
+                  new NoBrokersAvailableException(e.name, baseDtab, localDtab, limitedDtab))
             }
           }
         }
@@ -140,12 +141,12 @@ class BindingFactory[Req, Rep] private[naming] (
   }
 
   def apply(conn: ClientConnection): Future[Service[Req, Rep]] =
-    dtabCache(Dtabs(baseDtab(), Dtab.local), conn)
+    dtabCache(Dtabs(baseDtab(), Dtab.local, Dtab.limited), conn)
 
   def close(deadline: Time): Future[Unit] =
     Closable.sequence(dtabCache, nameTreeCache, nameCache).close(deadline)
 
-  override def status: Status = dtabCache.status(Dtabs(baseDtab(), Dtab.local))
+  override def status: Status = dtabCache.status(Dtabs(baseDtab(), Dtab.local, Dtab.limited))
 }
 
 object BindingFactory {
@@ -155,10 +156,11 @@ object BindingFactory {
   private[finagle] val NamerPathAnnotationKey = "clnt/namer.path"
   private[finagle] val DtabBaseAnnotationKey = "clnt/namer.dtab.base"
 
-  private case class Dtabs(base: Dtab, local: Dtab) extends CachedHashCode.ForClass {
+  private case class Dtabs(base: Dtab, local: Dtab, limited: Dtab) extends CachedHashCode.ForClass {
     override protected def computeHashCode: Int = {
       var hash = 1
       hash = 31 * hash + base.hashCode
+      hash = 31 * hash + limited.hashCode
       hash = 31 * hash + local.hashCode
       hash
     }
@@ -283,7 +285,7 @@ object BindingFactory {
         // resulting in wasteful connections. The second condition applies only if
         // EagerConnectionsType.ForceWithDtab is false.
         val finalParams =
-          if (!eagerlyConnect || (!forceWithDtab && !Dtab.local.isEmpty))
+          if (!eagerlyConnect || (!forceWithDtab && !(Dtab.local.isEmpty && Dtab.limited.isEmpty)))
             updatedParams + EagerConnections(false)
           else updatedParams
 
@@ -292,7 +294,7 @@ object BindingFactory {
       }
 
       val factory = dest match {
-        case bound @ Name.Bound(addr) => newStack(label, bound)
+        case bound @ Name.Bound(_) => newStack(label, bound)
 
         case Name.Path(path) =>
           val BaseDtab(baseDtab) = params[BaseDtab]