+++ CHANGELOG.rst
@@ -7,6 +7,16 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+New Features
+~~~~~~~~~~~~
+
+* finagle-core: Introduce `Dtab.limited`, which is a process-local `Dtab` that will
+  NOT be remotely broadcast for protocols for any protocol, where `Dtab.local` will
+  be broadcast for propagation on supported protocols. For path name resolution, the
+  `Dtab.local` will take precedence over the `Dtab.limited`, if the same path is
+  defined in both, and both take precedence over the `Dtab.base`. The existing
+  `Dtab.local` request propagation behavior remains unchanged. ``PHAB_ID=D677860``
+
 Runtime Behavior Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~